+++ lib/lightning/version_control/version_control.ex
@@ -8,9 +8,11 @@ defmodule Lightning.VersionControl do
 
   import Ecto.Query, warn: false
 
+  alias Ecto.Multi
   alias Lightning.Accounts.User
   alias Lightning.Extensions.UsageLimiting
   alias Lightning.Repo
+  alias Lightning.VersionControl.Audit
   alias Lightning.VersionControl.Events
   alias Lightning.VersionControl.GithubClient
   alias Lightning.VersionControl.GithubError
@@ -35,6 +37,10 @@ defmodule Lightning.VersionControl do
 
     Repo.transact(fn ->
       with {:ok, repo_connection} <- Repo.insert(changeset),
+           {:ok, _audit} <-
+             repo_connection
+             |> Audit.repo_connection(:created, user)
+             |> Repo.insert(),
            :ok <-
              VersionControlUsageLimiter.limit_github_sync(
                repo_connection.project_id
@@ -103,10 +109,15 @@ defmodule Lightning.VersionControl do
   Deletes a github connection
   """
   def remove_github_connection(repo_connection, user) do
-    repo_connection
-    |> Repo.delete()
+    Multi.new()
+    |> Multi.delete(:delete_repo_connection, repo_connection)
+    |> Multi.insert(
+      :audit,
+      Audit.repo_connection(repo_connection, :removed, user)
+    )
+    |> Repo.transaction()
     |> tap(fn
-      {:ok, repo_connection} ->
+      {:ok, %{delete_repo_connection: repo_connection}} ->
         undo_repo_actions(
           repo_connection,
           user